# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация API для сайта технического обслуживания автомобилей",
        "version": "1.0.0"
    },
    "paths": {
        "/Calculate": {
            "post": {
                "tags": [
                    "Calculate"
                ],
                "summary": "Рассчитать стоимость выбранных сервисов",
                "operationId": "servicesId",
                "requestBody": {
                    "description": "введите id необходимых сервисов, Response в первой строке, в самом начале(посчитанная сумма), остальное - Debugger, если включен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": [
                                        1,
                                        3,
                                        4
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Services/": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Получить список всех доступных сервисов",
                "description": "Получаем список всех доступных сервисов",
                "operationId": "getServices",
                "parameters": [
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Параметры с Get, для этого запроса пока не принимаются на бэке, это образец",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/TechServiceType": {
            "get": {
                "tags": [
                    "TechServiceType"
                ],
                "summary": "Получить список всех доступных категорий сервисов",
                "description": "Получаем список всех доступных категорий сервисов",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/TechServiceType/{id}": {
            "get": {
                "tags": [
                    "TechServiceType/id"
                ],
                "summary": "Получить список сервисов выбранной категории",
                "description": "Получаем список сервисов выбранной категории",
                "operationId": "servicesTypeId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "введите id необходимой категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Calculator": {
                "title": "Calculate",
                "description": "Рассчет стоимости выбранных сервисов",
                "xml": {
                    "name": "Calculate"
                }
            },
            "Service": {
                "title": "Services",
                "description": "Services model",
                "properties": {
                    "": {
                        "title": "Services",
                        "description": "Список сервисов",
                        "type": "object",
                        "example": "[{title: Замена передного тормозного диска, coefficient: 1, tech_service_types_id: 2},...,{}]"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Services"
                }
            },
            "TechServiceType": {
                "title": "TechServiceType",
                "description": "TechServiceType model",
                "properties": {
                    "": {
                        "title": "TechServiceType",
                        "description": "Список категорий сервисов",
                        "type": "object",
                        "example": "[{id: 1, title: ДВС&Трансмиссия},...,{}]"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TechServiceType"
                }
            }
        }
    }
}